backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
httpx==0.28.1
python-multipart==0.0.20
python-dotenv==1.0.1
sqlalchemy==2.0.36
alembic==1.14.0
aiomysql==0.2.0
aiosqlite==0.20.0
